Change __iter__ to return Nodes instead of strings
It was inconsistent for node['foo'] to return a Node, while iter(node)
would return strings. This changes that behavior so one can easily do
something like:

    for subnode in node:
        # use subnode

without having to resort to silliness like:

    for subnode in (node[name] for name in node):
        # use subnode

Updated (simplified) the example in the Usage too.

---

That change alone would make it difficult to get a Node's name. So this
also adds a __str__ implementation, which returns the basename() of the
path.